LibGfx: Move TTF files from TrueTypeFont/ to Font/TrueType/
