+++ Meta/Lagom/CMakeLists.txt
@@ -340,11 +340,11 @@ if (BUILD_LAGOM)
 
     # Gfx
     file(GLOB LIBGFX_SOURCES CONFIGURE_DEPENDS "../../Userland/Libraries/LibGfx/*.cpp")
-    file(GLOB LIBGFX_TTF_SOURCES CONFIGURE_DEPENDS "../../Userland/Libraries/LibGfx/TrueTypeFont/*.cpp")
+    file(GLOB LIBGFX_TTF_SOURCES CONFIGURE_DEPENDS "../../Userland/Libraries/LibGfx/Font/TrueType/*.cpp")
     lagom_lib(Gfx gfx
-        SOURCES ${LIBGFX_SOURCES} ${LIBGFX_TTF_SOURCES}
-        LIBS m LagomCompress LagomTextCodec LagomIPC
-    )
+            SOURCES ${LIBGFX_SOURCES} ${LIBGFX_TTF_SOURCES}
+            LIBS m LagomCompress LagomTextCodec LagomIPC
+            )
 
     # GPU
     file(GLOB LIBGPU_SOURCES CONFIGURE_DEPENDS "../../Userland/Libraries/LibGPU/*.cpp")
