+++ Userland/Libraries/LibGfx/Font/TrueType/Font.h
@@ -13,9 +13,9 @@
 #include <LibGfx/Bitmap.h>
 #include <LibGfx/Font.h>
 #include <LibGfx/Size.h>
-#include <LibGfx/TrueTypeFont/Cmap.h>
-#include <LibGfx/TrueTypeFont/Glyf.h>
-#include <LibGfx/TrueTypeFont/Tables.h>
+#include <LibGfx/Font/TrueType/Cmap.h>
+#include <LibGfx/Font/TrueType/Glyf.h>
+#include <LibGfx/Font/TrueType/Tables.h>
 
 #define POINTS_PER_INCH 72.0f
 #define DEFAULT_DPI 96
