+++ Userland/Libraries/LibGfx/Font/TrueType/Tables.h
