+++ Userland/Libraries/LibGfx/FontDatabase.cpp
@@ -10,7 +10,7 @@
 #include <LibCore/DirIterator.h>
 #include <LibGfx/Font.h>
 #include <LibGfx/FontDatabase.h>
-#include <LibGfx/TrueTypeFont/Font.h>
+#include <LibGfx/Font/TrueType/Font.h>
 #include <LibGfx/Typeface.h>
 #include <stdlib.h>
