+++ Userland/Libraries/LibGfx/Typeface.h
@@ -12,7 +12,7 @@
 #include <AK/Vector.h>
 #include <LibGfx/BitmapFont.h>
 #include <LibGfx/Font.h>
-#include <LibGfx/TrueTypeFont/Font.h>
+#include <LibGfx/Font/TrueType/Font.h>
 
 namespace Gfx {
